convert social embed controls to palette/ts
