+++ src/client/apps/edit/components/content/sections/social_embed/controls.tsx
@@ -1,0 +1,59 @@
+import { Input } from "@artsy/reaction/dist/Components/Input"
+import { SectionData } from "@artsy/reaction/dist/Components/Publishing/Typings"
+import {
+  onChangeSection,
+  removeSection,
+} from "client/actions/edit/sectionActions"
+import { FormLabel } from "client/components/form_label"
+import React, { Component } from "react"
+import { connect } from "react-redux"
+import SectionControls from "../../section_controls"
+
+interface SocialEmbedControlsProps {
+  onChangeSectionAction: (key: string, val: any) => void
+  removeSectionAction: (i: number) => void
+  section: SectionData
+  sectionIndex: number
+}
+
+export class SocialEmbedControls extends Component<SocialEmbedControlsProps> {
+  componentWillUnmount = () => {
+    const { removeSectionAction, section, sectionIndex } = this.props
+
+    if (!section.url) {
+      removeSectionAction(sectionIndex)
+    }
+  }
+
+  render() {
+    const { onChangeSectionAction, section } = this.props
+
+    return (
+      <SectionControls>
+        <FormLabel color="white">Social URL</FormLabel>
+        <Input
+          block
+          autoFocus
+          value={section.url || ""}
+          onChange={e => onChangeSectionAction("url", e.currentTarget.value)}
+          placeholder="https://www.instagram.com/"
+        />
+      </SectionControls>
+    )
+  }
+}
+
+const mapStateToProps = state => ({
+  sectionIndex: state.edit.sectionIndex,
+  section: state.edit.section,
+})
+
+const mapDispatchToProps = {
+  onChangeSectionAction: onChangeSection,
+  removeSectionAction: removeSection,
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(SocialEmbedControls)
